==> systematic/split_generation/0_split_generation_research.ipynb <==
import numpy as np
import pyarrow.parquet as pq
import pandas as pd
import pyarrow.compute as pc

parquet_file = pq.ParquetFile('../../train.parquet')
total_records = parquet_file.metadata.num_rows
# Optionally, you can inspect the schema to confirm field names and types
print(parquet_file.schema)

column_data = parquet_file.read(columns=['buildingblock1_smiles'])
unique_values = pc.unique(column_data.column(0))
unique_list1_train = unique_values.to_pylist()

column_data = parquet_file.read(columns=['buildingblock2_smiles'])
unique_values = pc.unique(column_data.column(0))
unique_list2_train = unique_values.to_pylist()

column_data = parquet_file.read(columns=['buildingblock3_smiles'])
unique_values = pc.unique(column_data.column(0))
unique_list3_train = unique_values.to_pylist()

import pickle
with open('../../unique_blocks.pkl','wb') as f:
    pickle.dump((unique_list1_train, unique_list2_train, unique_list3_train), f)

import random
random.shuffle(unique_list1_train)
split1_block1 = unique_list1_train[0:15] 
split2_block1 = unique_list1_train[15:30] 
split3_block1 = unique_list1_train[30:45] 
split4_block1 = unique_list1_train[45:60] 
split5_block1 = unique_list1_train[60:75] 

unique_list1_train

import random
block23 = list(set(unique_list2_train).union(set(unique_list3_train)))

random.shuffle(block23)
split1_block23 = block23[0:30] 
split2_block23 = block23[30:60] 
split3_block23 = block23[60:90] 
split4_block23 = block23[90:120] 
split5_block23 = block23[120:150] 

import pickle

with open('../../block_splits.pkl', 'wb') as f:
    pickle.dump(((split1_block1, split2_block1, split3_block1, split4_block1, split5_block1), (split1_block23, split2_block23, split3_block23, split4_block23, split5_block23)),f)

parquet_file = pq.ParquetFile('../../test.parquet')
total_records = parquet_file.metadata.num_rows
# Optionally, you can inspect the schema to confirm field names and types
print(parquet_file.schema)

column_data = parquet_file.read(columns=['buildingblock1_smiles'])
unique_values = pc.unique(column_data.column(0))
unique_list1_test = unique_values.to_pylist()

column_data = parquet_file.read(columns=['buildingblock2_smiles'])
unique_values = pc.unique(column_data.column(0))
unique_list2_test = unique_values.to_pylist()

column_data = parquet_file.read(columns=['buildingblock3_smiles'])
unique_values = pc.unique(column_data.column(0))
unique_list3_test = unique_values.to_pylist()

train_unique = set(unique_list1_train).union(set(unique_list2_train)).union(set(unique_list3_train) )
test_unique = set(unique_list1_test).union(set(unique_list2_test)).union(set(unique_list3_test) )

print(f'len train_unique_c1 {len(set(unique_list1_train))} train_unique_c2 {len(unique_list2_train)} train_unique_c3 {len(unique_list3_train)}')
print(f'len test_unique_c1 {len(unique_list1_test)} test_unique_c2 {len(unique_list2_test)} test_unique_c3 {len(unique_list3_test)}')


print(f'len delta test - train {len(set(unique_list1_test) - set(unique_list1_train))}')
print(f'len delta test - train {len(set(unique_list2_test) - set(unique_list2_train))}')
print(f'len delta test - train {len(set(unique_list3_test) - set(unique_list3_train))}')

print(f'len delta test - train {len(set(unique_list2_test).union(unique_list3_test) - set(unique_list2_train).union(unique_list3_train))}')
print(f'len delta test - train {len(set(unique_list3_test) - set(unique_list3_train))}')

len(set(unique_list2_test).union(unique_list3_test))

len(set(unique_list2_train).union(unique_list3_train))

len(set(unique_list1_test)-set(unique_list1_train))


print(f'len train_unique {len(train_unique)}')
print(f'len test unique {len(test_unique)}')
print(f'len delta train - test {len(train_unique - test_unique)}')
print(f'len delta test - train {len(test_unique - train_unique)}')

data_test = pd.read_parquet('../../test.parquet')

data_test.head()

len(data_test)

data_test[((data_test['buildingblock1_smiles'].isin(unique_list1_train)) | (data_test['buildingblock2_smiles'].isin(unique_list2_train) | data_test['buildingblock2_smiles'].isin(unique_list3_train))  | (data_test['buildingblock3_smiles'].isin(unique_list3_train)))]

train_unique[train_unique['buildingblock3_smiles'].isin(unique_list1_train)]

import random
u1_random = random.sample(unique_list1_train, 10)
u2_random = random.sample(unique_list2_train, 20)
u3_random = random.sample(unique_list3_train, 20)

parquet_file = pq.ParquetFile('../../train.parquet')

from tqdm.auto import tqdm
x_inside = 0
x_inside_i = 0
drop_i = 0
x_rest = 0
total = 0
overlapping = set(u1_random).union(set(u2_random).union(u3_random))
for x in tqdm(parquet_file.iter_batches(batch_size=1000000)):
    x = x.to_pandas()
    x_inside += len(x[x['buildingblock1_smiles'].isin(u1_random) & x['buildingblock2_smiles'].isin(u2_random) & x['buildingblock3_smiles'].isin(u3_random)])
    x_inside_i += len(x[x['buildingblock1_smiles'].isin(u1_random) & x['buildingblock3_smiles'].isin(u3_random) & x['buildingblock2_smiles'].isin(u2_random)]) 
    drop_i += len(x[x['buildingblock1_smiles'].isin(u1_random) | x['buildingblock2_smiles'].isin(overlapping) | x['buildingblock3_smiles'].isin(overlapping)])  

    x_rest += len(x)-x_inside-x_inside_i
    total += len(x)
    

total

x_inside

x_inside_i

drop_i

x_rest

def generate_train(file, unique1, unique_rest):
    parquet_file = pq.ParquetFile('../../train.parquet')
    train_ids = []
    test_ids = []
    sum_ids = set(unique1).union(set(unique_rest))
    for x in tqdm(parquet_file.iter_batches(batch_size=1000000)):
        x = x.to_pandas()
        x_test = x['buildingblock1_smiles'].isin(unique1) & x['buildingblock2_smiles'].isin(unique_rest) & x['buildingblock3_smiles'].isin(unique_rest)
        x_train = ~(x['buildingblock1_smiles'].isin(sum_ids) | x['buildingblock2_smiles'].isin(sum_ids) | x['buildingblock3_smiles'].isin(sum_ids))

        x = x.drop(['buildingblock1_smiles','buildingblock2_smiles','buildingblock3_smiles','molecule_smiles','protein_name','binds'],axis=1)
        x['is_train'] = x_train
        x['is_test_unique'] = x_test

def generate_split(file, unique1, unique_rest):
    parquet_file = pq.ParquetFile('../../train.parquet')
    train_ids = []
    test_ids = []
    sum_ids = set(unique1).union(set(unique_rest))
    for x in tqdm(parquet_file.iter_batches(batch_size=1000000)):
        x = x.to_pandas()
        x_test = x['buildingblock1_smiles'].isin(unique1) & x['buildingblock2_smiles'].isin(unique_rest) & x['buildingblock3_smiles'].isin(unique_rest)
        x_train = ~(x['buildingblock1_smiles'].isin(sum_ids) | x['buildingblock2_smiles'].isin(sum_ids) | x['buildingblock3_smiles'].isin(sum_ids))

        x = x.drop(['buildingblock1_smiles','buildingblock2_smiles','buildingblock3_smiles','molecule_smiles','protein_name','binds'],axis=1)
        x['is_train'] = x_train
        x['is_test_unique'] = x_test
        

    

def protein_name_eda(dataset_path):
    protein_name =  pd.read_parquet(dataset_path, engine = 'pyarrow', columns=['protein_name']).protein_name.to_numpy()
    protein_name_reshaped = np.reshape(protein_name, [-1, 3])
    print(np.mean(protein_name_reshaped[:, 0] == 'BRD4'))
    print(np.mean(protein_name_reshaped[:, 1] == 'HSA'))
    print(np.mean(protein_name_reshaped[:, 2] == 'sEH'))
    
protein_name_eda('../../train.parquet')

DEBUG = False
if DEBUG:
    NUM_ROWS = 30000000
else:
    NUM_ROWS = 295246830

def get_unique_BB(dataset_path, col):
    BBs = pd.read_parquet(dataset_path, engine = 'pyarrow', columns=[col])
    BBs = BBs[:NUM_ROWS]
    BBs = BBs.to_numpy()[:, 0]
    BBs_reshaped = np.reshape(BBs, [-1, 3])
    
    if np.mean(BBs_reshaped[:, 0] == BBs_reshaped[:, 1]) != 1:
        print('ERROR')
    if np.mean(BBs_reshaped[:, 0] == BBs_reshaped[:, 2]) != 1:
        print('ERROR')
    
    BBs_unique = np.unique(BBs_reshaped[:, 0])
    BBs_unique = list(BBs_unique)
    BBs_dict = {BBs_unique[i]:i for i in range(len(BBs_unique))}
    BBs_dict_reverse = {i:BBs_unique[i] for i in range(len(BBs_unique))}
    return BBs_dict, BBs_dict_reverse

BBs_dict_1, BBs_dict_reverse_1 = get_unique_BB('../../train.parquet', 'buildingblock1_smiles')
print(len(BBs_dict_1))
BBs_dict_2, BBs_dict_reverse_2 = get_unique_BB('../../train.parquet', 'buildingblock2_smiles')
print(len(BBs_dict_2))
BBs_dict_3, BBs_dict_reverse_3 = get_unique_BB('../../train.parquet', 'buildingblock3_smiles')
print(len(BBs_dict_3))
