==> systematic/split_generation/1_distinct_islands.ipynb <==
import numpy as np
import pyarrow.parquet as pq
import pandas as pd
import pyarrow.compute as pc
from tqdm.auto import tqdm
import cv2
from multiprocessing import Pool

parquet_file = pq.ParquetFile('../../train.parquet')
total_records = parquet_file.metadata.num_rows
# Optionally, you can inspect the schema to confirm field names and types
print(parquet_file.schema)

column_data = parquet_file.read(columns=['buildingblock1_smiles'])
unique_values = pc.unique(column_data.column(0))
unique_list1_train = unique_values.to_pylist()

column_data = parquet_file.read(columns=['buildingblock2_smiles'])
unique_values = pc.unique(column_data.column(0))
unique_list2_train = unique_values.to_pylist()

column_data = parquet_file.read(columns=['buildingblock3_smiles'])
unique_values = pc.unique(column_data.column(0))
unique_list3_train = unique_values.to_pylist()

import pickle
with open('../spl1.pkl','wb') as f:
    pickle.dump(unique_list1_train,f)
with open('../spl2.pkl','wb') as f:
    pickle.dump(unique_list2_train,f)
with open('../spl3.pkl','wb') as f:
    pickle.dump(unique_list3_train,f)

blocks = list(set(unique_list2_train).union(set(unique_list3_train).union(set(unique_list1_train))))

len(blocks)

blocks_to_i = {x: i for i, x in enumerate(blocks)}

parquet_file = pq.ParquetFile('../../test.parquet')

del column_data

def process_split(x):
    x = x.to_pandas()
    graph = np.zeros((2110,2110), dtype=bool)
    
    for k, v in x.iterrows():       
        graph[blocks_to_i[v['buildingblock1_smiles']],blocks_to_i[v['buildingblock1_smiles']]] = True
        graph[blocks_to_i[v['buildingblock2_smiles']],blocks_to_i[v['buildingblock2_smiles']]] = True
        graph[blocks_to_i[v['buildingblock3_smiles']],blocks_to_i[v['buildingblock3_smiles']]] = True

        graph[blocks_to_i[v['buildingblock1_smiles']],blocks_to_i[v['buildingblock2_smiles']]] = True
        graph[blocks_to_i[v['buildingblock2_smiles']],blocks_to_i[v['buildingblock1_smiles']]] = True

        graph[blocks_to_i[v['buildingblock1_smiles']],blocks_to_i[v['buildingblock3_smiles']]] = True
        graph[blocks_to_i[v['buildingblock3_smiles']],blocks_to_i[v['buildingblock1_smiles']]] = True

        graph[blocks_to_i[v['buildingblock2_smiles']],blocks_to_i[v['buildingblock3_smiles']]] = True
        graph[blocks_to_i[v['buildingblock3_smiles']],blocks_to_i[v['buildingblock2_smiles']]] = True
    return graph

BATCH = 100_000
total_records = parquet_file.metadata.num_rows
total_batches = (total_records + BATCH - 1) // BATCH  # Ceiling division  
with Pool(processes=30) as pool:  # Adjust the number of processes according to your system
        # Use imap_unordered to process data as it is read
        results = list(tqdm(pool.imap_unordered(process_split, parquet_file.iter_batches(batch_size=BATCH)),
                            total=total_batches, desc="Processing batches"))

results_stack = np.stack(results)

results_stack.shape

final = results_stack.any(axis=0)

import numpy as np
from scipy.sparse import csr_matrix
from scipy.sparse.csgraph import connected_components

sparse_graph = csr_matrix(final.astype(int))

n_components, labels = connected_components(csgraph=sparse_graph, directed=False, return_labels=True)

print(f"Number of connected components: {n_components}")

np.unique(labels, return_counts=True)

len(labels)

TEST_MAIN = '/home/anonymous/belka/test.parquet'
test = pd.read_parquet(TEST_MAIN)

def get_score(row):
    return labels[blocks_to_i[row['buildingblock1_smiles']]]
test['island'] = test.apply(get_score, axis=1)


test.to_parquet('/home/anonymous/belka/test_with_islands.parquet',index=False)

sum((test['island']==) == ((test['buildingblock3_smiles'].isin(set(blocks))) | (test['buildingblock2_smiles'].isin(set(blocks))) | (test['buildingblock1_smiles'].isin(set(blocks)))))

# import pickle
# with open('test_connected.pickle','wb') as f:
#     pickle.dump((labels, np.unique(labels, return_counts=True), blocks_to_i), f)

import networkx as nx

graph = nx.Graph(sparse_graph)
cut_value, (set1, set2) = nx.stoer_wagner(graph)

stats
