==> systematic/split_generation/3_split_generation_new.ipynb <==
import numpy as np
import pandas as pd
from pyarrow.parquet import ParquetFile
import pickle
import os

train_data_islands = pd.read_parquet('/home/anonymous/belka/test_with_islands.parquet')

train_data_islands

molecule_smiles_test_island1 = train_data_islands[train_data_islands.island==1].molecule_smiles.unique()
molecule_smiles_test_island23 = train_data_islands[(train_data_islands.island==2) | (train_data_islands.island==3)].molecule_smiles.unique()


test_data_original = pd.read_parquet('../test_data/test.parquet')

test_data_original[test_data_original.molecule_smiles.isin(molecule_smiles_test_island1)].sample(frac=1).to_parquet('../test_data/test_shuffled_is1.parquet', index = False)

test_data_original[test_data_original.molecule_smiles.isin(molecule_smiles_test_island23)].sample(frac=1).to_parquet('../test_data/test_shuffled_is23.parquet', index = False)

train_data_original = train_data_original[train_data_original.molecule_smiles.isin(molecule_smiles)]

train_data_original = train_data_original.sample(frac=1)

train_data_original

b1_unique = train_data_original['buildingblock1_smiles'].unique()
b2_unique = train_data_original['buildingblock2_smiles'].unique()
b3_unique = train_data_original['buildingblock3_smiles'].unique()

import random

b1_sample = random.sample(list(b1_unique), 27)
b2_sample = random.sample(list(b2_unique), 69)
b3_sample = random.sample(list(b3_unique), 87)

train_data = train_data_original.copy()
mask_unique = train_data['buildingblock1_smiles'].isin(b1_sample) & train_data['buildingblock2_smiles'].isin(b2_sample) & train_data['buildingblock3_smiles'].isin(b3_sample) 
test_unique = train_data[mask_unique]
train_data = train_data[~mask_unique]
left_over_mask = train_data['buildingblock1_smiles'].isin(b1_sample) | train_data['buildingblock2_smiles'].isin(b2_sample) | train_data['buildingblock3_smiles'].isin(b3_sample) 
train_data = train_data[~left_over_mask]
from sklearn.model_selection import train_test_split
train, test = train_test_split(train_data, test_size=0.1)
test_unique.to_parquet('../train_data/splits/s1_unique.parquet', index = False)
train.to_parquet('../train_data/splits/s1_train.parquet', index = False)
test.to_parquet('../train_data/splits/s1_random.parquet', index = False)

train_data = train_data_original.copy()
b1_sample = random.sample(list(b1_unique), 27)
b2_sample = random.sample(list(b2_unique), 69)
b3_sample = random.sample(list(b3_unique), 87)
mask_unique = train_data['buildingblock1_smiles'].isin(b1_sample) & train_data['buildingblock2_smiles'].isin(b2_sample) & train_data['buildingblock3_smiles'].isin(b3_sample) 
test_unique = train_data[mask_unique]
train_data = train_data[~mask_unique]
left_over_mask = train_data['buildingblock1_smiles'].isin(b1_sample) | train_data['buildingblock2_smiles'].isin(b2_sample) | train_data['buildingblock3_smiles'].isin(b3_sample) 
train_data = train_data[~left_over_mask]
from sklearn.model_selection import train_test_split
train, test = train_test_split(train_data, test_size=0.1)
test_unique.to_parquet('../train_data/splits/s2_unique.parquet', index = False)
train.to_parquet('../train_data/splits/s2_train.parquet', index = False)
test.to_parquet('../train_data/splits/s2_random.parquet', index = False)

train_data = train_data_original.copy()
b1_sample = random.sample(list(b1_unique), 27)
b2_sample = random.sample(list(b2_unique), 69)
b3_sample = random.sample(list(b3_unique), 87)
mask_unique = train_data['buildingblock1_smiles'].isin(b1_sample) & train_data['buildingblock2_smiles'].isin(b2_sample) & train_data['buildingblock3_smiles'].isin(b3_sample) 
test_unique = train_data[mask_unique]
train_data = train_data[~mask_unique]
left_over_mask = train_data['buildingblock1_smiles'].isin(b1_sample) | train_data['buildingblock2_smiles'].isin(b2_sample) | train_data['buildingblock3_smiles'].isin(b3_sample) 
train_data = train_data[~left_over_mask]
from sklearn.model_selection import train_test_split
train, test = train_test_split(train_data, test_size=0.1)
test_unique.to_parquet('../train_data/splits/s3_unique.parquet', index = False)
train.to_parquet('../train_data/splits/s3_train.parquet', index = False)
test.to_parquet('../train_data/splits/s3_random.parquet', index = False)

len(test)
